+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `mouc convert-format` now preserves YAML block scalars and list formatting
+
 ## [0.7.3] - 2025-12-19
 
 ### Changed